chore(workspace,clippy): batch 2 — 29 more autofix wins across 10 files
cargo clippy --fix --allow-dirty --workspace --tests (2nd pass):

- consensus/state_sync.rs (7 fixes) — needless slice clones, redundant
  closure captures
- consensus/tendermint.rs (8 fixes total in 2 passes) — same patterns
- consensus/validator_set.rs (3 fixes)
- consensus/audit_tests.rs (2 fixes) — test cleanup
- consensus/mempool.rs (1 fix)
- consensus/tests/byzantine_adversarial.rs (2 fixes)
- consensus/tests/mcc_phase_d.rs (1 fix)
- crypto/verkle.rs (1 fix) — needless 'as u8' cast on a u8 value
- mev-detect/lib.rs (1 fix)
- proving/lib.rs — &vec![0u8; N] -> &[0u8; N] (array literal slice)

Workspace warning total: 81 -> 52 across the 2-pass sweep.

716/716 consensus lib tests still pass.

+++ crates/evaporchain-crypto/src/verkle.rs
@@ -860,7 +860,7 @@ mod tests {
             );
             // Wrong-value tampering also rejected at deep paths.
             let mut bad_proof = proof.clone();
-            bad_proof.value = Some(make_value(((val[0]).wrapping_add(1)) as u8));
+            bad_proof.value = Some(make_value(((val[0]).wrapping_add(1))));
             assert!(
                 !VerkleTrie::verify(&bad_proof, &root),
                 "tampered-value proof must reject at deep path"
